Modifica el flow de alta de pedido, y de facturación.
