+++ src/main/java/sic/vista/swing/NuevaFacturaVentaGUI.java
@@ -8,12 +8,8 @@ import java.beans.PropertyVetoException;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.text.ParseException;
-import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.GregorianCalendar;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.ResourceBundle;
@@ -37,11 +33,9 @@ import sic.modelo.RenglonFactura;
 import sic.modelo.UsuarioActivo;
 import sic.modelo.NuevoRenglonFactura;
 import sic.modelo.FormaDePago;
-import sic.modelo.Movimiento;
 import sic.modelo.NuevaFacturaVenta;
 import sic.modelo.NuevosResultadosComprobante;
 import sic.modelo.Pedido;
-import sic.modelo.Producto;
 import sic.modelo.ProductoFaltante;
 import sic.modelo.ProductosParaVerificarStock;
 import sic.modelo.Resultados;
@@ -49,7 +43,6 @@ import sic.modelo.Rol;
 import sic.modelo.TipoDeComprobante;
 import sic.modelo.Transportista;
 import sic.util.DecimalesRenderer;
-import sic.util.Utilidades;
 
 public class NuevaFacturaVentaGUI extends JInternalFrame {
 
@@ -61,25 +54,14 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
     private final HotKeysHandler keyHandler = new HotKeysHandler();
     private final Logger LOGGER = LoggerFactory.getLogger(this.getClass());
     private final Dimension sizeInternalFrame = new Dimension(1200, 700);    
-    private int cantidadMaximaRenglones = 0;
     private final List<Rol> rolesDeUsuario = UsuarioActivo.getInstance().getUsuario().getRoles();
     private final static BigDecimal CIEN = new BigDecimal("100");
-
-    public NuevaFacturaVentaGUI() {
-        this.initComponents();
-        this.pedido = null;        
-        ImageIcon iconoNoMarcado = new ImageIcon(getClass().getResource("/sic/icons/chkNoMarcado_16x16.png"));
-        this.tbtn_marcarDesmarcar.setIcon(iconoNoMarcado);        
-        dc_fechaVencimiento.setDate(new Date());
-        this.setListeners();
-    }   
     
     public NuevaFacturaVentaGUI(Pedido pedido) {
         this.initComponents();        
         this.pedido = pedido;
         ImageIcon iconoNoMarcado = new ImageIcon(getClass().getResource("/sic/icons/chkNoMarcado_16x16.png"));
         this.tbtn_marcarDesmarcar.setIcon(iconoNoMarcado);        
-        dc_fechaVencimiento.setDate(new Date());
         this.setListeners();
     }
     
@@ -92,16 +74,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
                 .recargoPorcentaje(new BigDecimal(txt_Recargo_porcentaje.getValue().toString()))
                 .descuentoPorcentaje(new BigDecimal(txt_Descuento_porcentaje.getValue().toString()))
                 .build();
-        if (this.dc_fechaVencimiento.getDate() != null) {
-            Calendar cal = new GregorianCalendar();
-            cal.setTime(this.dc_fechaVencimiento.getDate());
-            cal.set(Calendar.HOUR_OF_DAY, 23);
-            cal.set(Calendar.MINUTE, 59);
-            cal.set(Calendar.SECOND, 58);
-            nuevaFacturaVenta.setFechaVencimiento(cal.getTime().toInstant()
-                    .atZone(ZoneId.systemDefault())
-                    .toLocalDate());
-        }
         List<NuevoRenglonFactura> nuevosRenglones = new ArrayList<>();
         this.renglonesFactura.forEach(renglon -> {
             NuevoRenglonFactura renglonNuevo = NuevoRenglonFactura.builder()
@@ -110,25 +82,16 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
                     .build();
             nuevosRenglones.add(renglonNuevo);
         });
-        nuevaFacturaVenta.setRenglones(nuevosRenglones);
         return nuevaFacturaVenta;
     } 
     
     private void setListeners() {
         cmb_TipoComprobante.addKeyListener(keyHandler);
-        btn_NuevoCliente.addKeyListener(keyHandler);
-        btn_BuscarCliente.addKeyListener(keyHandler);
-        btn_BuscarProductos.addKeyListener(keyHandler);
-        btn_QuitarProducto.addKeyListener(keyHandler);
         tbl_Resultado.addKeyListener(keyHandler);
-        txt_CodigoProducto.addKeyListener(keyHandler);
-        btn_BuscarPorCodigoProducto.addKeyListener(keyHandler);
         txt_Descuento_porcentaje.addKeyListener(keyHandler);
         txt_Recargo_porcentaje.addKeyListener(keyHandler);
         btn_Continuar.addKeyListener(keyHandler);
         tbtn_marcarDesmarcar.addKeyListener(keyHandler);        
-        dc_fechaVencimiento.addKeyListener(keyHandler);     
-        btnModificarCliente.addKeyListener(keyHandler); 
     }
     
     private void cargarPedidoParaFacturar() {
@@ -321,90 +284,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
         }
         tbl_Resultado.setModel(modeloTablaResultados);
     }
-
-    private void buscarProductoConVentanaAuxiliar() {
-        if (cantidadMaximaRenglones > renglonesFactura.size()) {
-            BuscarProductosGUI buscarProductosGUI = new BuscarProductosGUI(renglonesFactura, this.tipoDeComprobante, Movimiento.VENTA);
-            buscarProductosGUI.setModal(true);
-            buscarProductosGUI.setLocationRelativeTo(this);
-            buscarProductosGUI.setVisible(true);
-            if (buscarProductosGUI.debeCargarRenglon()) {
-                boolean renglonCargado = false;
-                for (RenglonFactura renglon : renglonesFactura) {
-                    if (renglon.getIdProductoItem() == buscarProductosGUI.getRenglonFactura().getIdProducto()) {
-                        renglonCargado = true;
-                    }
-                }
-                this.agregarRenglon(buscarProductosGUI.getRenglonFactura());
-                /*Si la tabla no contiene renglones, despues de agregar el renglon
-                 a la coleccion, carga el arreglo con los estados con un solo elemento, 
-                 cuyo valor es "Desmarcado" para evitar un nulo.*/
-                EstadoRenglon[] estadosRenglones = new EstadoRenglon[renglonesFactura.size()];
-                if (tbl_Resultado.getRowCount() == 0) {
-                    estadosRenglones[0] = EstadoRenglon.DESMARCADO;
-                } else {
-                    this.cargarEstadoDeLosChkEnTabla(tbl_Resultado, estadosRenglones);
-                    //Se ejecuta o no segun si el renglon ya existe. Si ya existe, no se ejecuta
-                    if (!renglonCargado) {
-                        estadosRenglones[tbl_Resultado.getRowCount()] = EstadoRenglon.DESMARCADO;
-                    }
-                }
-                this.cargarRenglonesAlTable(estadosRenglones);
-                this.calcularResultados();
-            }
-        } else {
-            JOptionPane.showMessageDialog(this,
-                    ResourceBundle.getBundle("Mensajes").getString("mensaje_maxima_cantidad_de_renglones"),
-                    "Error", JOptionPane.ERROR_MESSAGE);
-        }
-    }
-
-    private void buscarProductoPorCodigo() {
-        try {
-            Producto producto = RestClient.getRestTemplate().getForObject("/productos/busqueda?"
-                    + "codigo=" + txt_CodigoProducto.getText().trim(), Producto.class);
-            if (producto == null) {
-                JOptionPane.showMessageDialog(this, ResourceBundle.getBundle("Mensajes")
-                        .getString("mensaje_producto_no_encontrado"), "Error", JOptionPane.ERROR_MESSAGE);
-            } else {
-                NuevoRenglonFactura nuevoRenglon = NuevoRenglonFactura.builder()
-                        .cantidad(BigDecimal.ONE)
-                        .idProducto(producto.getIdProducto())
-                        .build();
-                boolean esValido = true;
-                List<ProductoFaltante> faltantes;
-                if (cmb_TipoComprobante.getSelectedItem() != TipoDeComprobante.PEDIDO) {
-                    faltantes = this.getProductosSinStockDisponible(Arrays.asList(nuevoRenglon));
-                    if (!faltantes.isEmpty()) {
-                        esValido = false;
-                        JOptionPane.showMessageDialog(this, ResourceBundle.getBundle("Mensajes")
-                                .getString("mensaje_producto_sin_stock_suficiente"), "Error", JOptionPane.ERROR_MESSAGE);
-                    }
-                }
-                if (esValido) {
-                    boolean renglonCargado = this.agregarRenglon(nuevoRenglon);
-                    EstadoRenglon[] estadosRenglones = new EstadoRenglon[renglonesFactura.size()];
-                    if (tbl_Resultado.getRowCount() == 0) {
-                        estadosRenglones[0] = EstadoRenglon.DESMARCADO;
-                    } else {
-                        this.cargarEstadoDeLosChkEnTabla(tbl_Resultado, estadosRenglones);
-                        if (!renglonCargado) {
-                            estadosRenglones[tbl_Resultado.getRowCount()] = EstadoRenglon.DESMARCADO;
-                        }
-                    }
-                    this.cargarRenglonesAlTable(estadosRenglones);
-                    this.calcularResultados();
-                    txt_CodigoProducto.setText("");
-                }
-            }
-        } catch (RestClientResponseException ex) {
-            JOptionPane.showMessageDialog(this, ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
-        } catch (ResourceAccessException ex) {
-            LOGGER.error(ex.getMessage());
-            JOptionPane.showMessageDialog(this, ResourceBundle.getBundle("Mensajes")
-                    .getString("mensaje_error_conexion"), "Error", JOptionPane.ERROR_MESSAGE);
-        }
-    }
     
     private void validarComponentesDeResultados() {
         if (txt_Descuento_porcentaje.isEditValid()) {
@@ -529,7 +408,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
             cantidades[i] = nuevosRenglonesFactura.get(i).getCantidad();
         }
         ProductosParaVerificarStock productosParaVerificarStock = ProductosParaVerificarStock.builder()
-                //.idSucursal(SucursalActiva.getInstance().getSucursal().getIdSucursal())
                 .cantidad(cantidades)
                 .idProducto(idsProductos)
                 .build();
@@ -544,31 +422,7 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
 
         @Override
         public void keyPressed(KeyEvent evt) {
-            if (evt.getKeyCode() == KeyEvent.VK_F2) {
-                btn_BuscarClienteActionPerformed(null);
-            }
-            
-            if (evt.getKeyCode() == KeyEvent.VK_F4) {
-                btn_BuscarProductosActionPerformed(null);
-            }
-            
-            if (evt.getKeyCode() == KeyEvent.VK_F5) {
-                btn_NuevoClienteActionPerformed(null);
-            }
-            
-            if (evt.getKeyCode() == KeyEvent.VK_F6) {
-                btnModificarClienteActionPerformed(null);
-            }
-
-            if (evt.getKeyCode() == KeyEvent.VK_F9) {
-                btn_ContinuarActionPerformed(null);
-            }
-
-            if (evt.getSource() == tbl_Resultado && evt.getKeyCode() == 127) {
-                btn_QuitarProductoActionPerformed(null);
-            }
-
-            if (evt.getSource() == tbl_Resultado && evt.getKeyCode() == KeyEvent.VK_TAB) {                
+            if (evt.getSource() == tbl_Resultado && evt.getKeyCode() == KeyEvent.VK_TAB) {
                 txt_Descuento_porcentaje.requestFocus();
             }
         }
@@ -579,25 +433,9 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
     private void initComponents() {
 
         panelGeneral = new javax.swing.JPanel();
-        panelCliente = new javax.swing.JPanel();
-        lblNombreCliente = new javax.swing.JLabel();
-        lblUbicacionCliente = new javax.swing.JLabel();
-        lbl_IDFiscalCliente = new javax.swing.JLabel();
-        lbl_CondicionIVACliente = new javax.swing.JLabel();
-        txt_CondicionIVACliente = new javax.swing.JTextField();
-        txtUbicacionCliente = new javax.swing.JTextField();
-        txtNombreCliente = new javax.swing.JTextField();
-        lblMontoCompraMinima = new javax.swing.JLabel();
-        txtIdFiscalCliente = new javax.swing.JTextField();
-        txtMontoCompraMinima = new javax.swing.JTextField();
         panelRenglones = new javax.swing.JPanel();
         sp_Resultado = new javax.swing.JScrollPane();
         tbl_Resultado = new javax.swing.JTable();
-        btn_BuscarProductos = new javax.swing.JButton();
-        btn_QuitarProducto = new javax.swing.JButton();
-        txt_CodigoProducto = new javax.swing.JTextField();
-        btn_BuscarPorCodigoProducto = new javax.swing.JButton();
-        tbtn_marcarDesmarcar = new javax.swing.JToggleButton();
         panelObservaciones = new javax.swing.JPanel();
         lbl_Observaciones = new javax.swing.JLabel();
         btn_AddComment = new javax.swing.JButton();
@@ -624,121 +462,45 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
         txt_Recargo_porcentaje = new javax.swing.JFormattedTextField();
         btn_Continuar = new javax.swing.JButton();
         panelEncabezado = new javax.swing.JPanel();
-        lbl_fechaDeVencimiento = new javax.swing.JLabel();
-        dc_fechaVencimiento = new com.toedter.calendar.JDateChooser();
         lbl_TipoDeComprobante = new javax.swing.JLabel();
         cmb_TipoComprobante = new javax.swing.JComboBox();
-        btn_NuevoCliente = new javax.swing.JButton();
-        btn_BuscarCliente = new javax.swing.JButton();
         lblSeparadorIzquierdo = new javax.swing.JLabel();
         lblSeparadorDerecho = new javax.swing.JLabel();
-        btnModificarCliente = new javax.swing.JButton();
+        panelCliente = new javax.swing.JPanel();
+        lblNombreCliente = new javax.swing.JLabel();
+        lblUbicacionCliente = new javax.swing.JLabel();
+        lbl_IDFiscalCliente = new javax.swing.JLabel();
+        lbl_CondicionIVACliente = new javax.swing.JLabel();
+        txt_CondicionIVACliente = new javax.swing.JTextField();
+        txtUbicacionCliente = new javax.swing.JTextField();
+        txtNombreCliente = new javax.swing.JTextField();
+        lblMontoCompraMinima = new javax.swing.JLabel();
+        txtIdFiscalCliente = new javax.swing.JTextField();
+        txtMontoCompraMinima = new javax.swing.JTextField();
+        tbtn_marcarDesmarcar = new javax.swing.JToggleButton();
 
         setResizable(true);
         setTitle("Nueva Factura de Venta");
         setFrameIcon(new javax.swing.ImageIcon(getClass().getResource("/sic/icons/SIC_16_square.png"))); // NOI18N
         addInternalFrameListener(new javax.swing.event.InternalFrameListener() {
-            public void internalFrameOpened(javax.swing.event.InternalFrameEvent evt) {
-                formInternalFrameOpened(evt);
-            }
-            public void internalFrameClosing(javax.swing.event.InternalFrameEvent evt) {
+            public void internalFrameActivated(javax.swing.event.InternalFrameEvent evt) {
             }
             public void internalFrameClosed(javax.swing.event.InternalFrameEvent evt) {
             }
-            public void internalFrameIconified(javax.swing.event.InternalFrameEvent evt) {
+            public void internalFrameClosing(javax.swing.event.InternalFrameEvent evt) {
+            }
+            public void internalFrameDeactivated(javax.swing.event.InternalFrameEvent evt) {
             }
             public void internalFrameDeiconified(javax.swing.event.InternalFrameEvent evt) {
             }
-            public void internalFrameActivated(javax.swing.event.InternalFrameEvent evt) {
+            public void internalFrameIconified(javax.swing.event.InternalFrameEvent evt) {
             }
-            public void internalFrameDeactivated(javax.swing.event.InternalFrameEvent evt) {
+            public void internalFrameOpened(javax.swing.event.InternalFrameEvent evt) {
+                formInternalFrameOpened(evt);
             }
         });
 
         panelGeneral.setBorder(javax.swing.BorderFactory.createEtchedBorder());
-
-        lblNombreCliente.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        lblNombreCliente.setText("Nombre:");
-
-        lblUbicacionCliente.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        lblUbicacionCliente.setText("Ubicación:");
-
-        lbl_IDFiscalCliente.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lbl_IDFiscalCliente.setText("CUIT o DNI:");
-
-        lbl_CondicionIVACliente.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        lbl_CondicionIVACliente.setText("Categoria IVA:");
-
-        txt_CondicionIVACliente.setEditable(false);
-        txt_CondicionIVACliente.setFocusable(false);
-
-        txtUbicacionCliente.setEditable(false);
-        txtUbicacionCliente.setFocusable(false);
-
-        txtNombreCliente.setEditable(false);
-        txtNombreCliente.setFocusable(false);
-
-        lblMontoCompraMinima.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lblMontoCompraMinima.setText("Compra Mínima $:");
-
-        txtIdFiscalCliente.setEditable(false);
-        txtIdFiscalCliente.setFocusable(false);
-
-        txtMontoCompraMinima.setEditable(false);
-
-        javax.swing.GroupLayout panelClienteLayout = new javax.swing.GroupLayout(panelCliente);
-        panelCliente.setLayout(panelClienteLayout);
-        panelClienteLayout.setHorizontalGroup(
-            panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelClienteLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                    .addComponent(lblNombreCliente, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(lblUbicacionCliente, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(lbl_CondicionIVACliente))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelClienteLayout.createSequentialGroup()
-                        .addComponent(txt_CondicionIVACliente, javax.swing.GroupLayout.PREFERRED_SIZE, 581, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(lbl_IDFiscalCliente)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(txtIdFiscalCliente))
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelClienteLayout.createSequentialGroup()
-                        .addComponent(txtNombreCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 635, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(lblMontoCompraMinima)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(txtMontoCompraMinima))
-                    .addComponent(txtUbicacionCliente, javax.swing.GroupLayout.Alignment.TRAILING))
-                .addGap(0, 0, 0))
-        );
-
-        panelClienteLayout.linkSize(javax.swing.SwingConstants.HORIZONTAL, new java.awt.Component[] {txtNombreCliente, txt_CondicionIVACliente});
-
-        panelClienteLayout.setVerticalGroup(
-            panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelClienteLayout.createSequentialGroup()
-                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
-                    .addComponent(lblMontoCompraMinima)
-                    .addComponent(txtNombreCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblNombreCliente)
-                    .addComponent(txtMontoCompraMinima, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(txtUbicacionCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lblUbicacionCliente))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
-                    .addComponent(txtIdFiscalCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(txt_CondicionIVACliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(lbl_IDFiscalCliente)
-                    .addComponent(lbl_CondicionIVACliente)))
-        );
-
-        panelClienteLayout.linkSize(javax.swing.SwingConstants.VERTICAL, new java.awt.Component[] {txtIdFiscalCliente, txtNombreCliente, txtUbicacionCliente, txt_CondicionIVACliente});
-
-        panelClienteLayout.linkSize(javax.swing.SwingConstants.VERTICAL, new java.awt.Component[] {lblMontoCompraMinima, lbl_IDFiscalCliente});
 
         tbl_Resultado.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
@@ -761,86 +523,16 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
         });
         sp_Resultado.setViewportView(tbl_Resultado);
 
-        btn_BuscarProductos.setForeground(java.awt.Color.blue);
-        btn_BuscarProductos.setIcon(new javax.swing.ImageIcon(getClass().getResource("/sic/icons/Product_16x16.png"))); // NOI18N
-        btn_BuscarProductos.setText("Buscar Producto (F4)");
-        btn_BuscarProductos.setFocusable(false);
-        btn_BuscarProductos.setPreferredSize(new java.awt.Dimension(200, 30));
-        btn_BuscarProductos.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_BuscarProductosActionPerformed(evt);
-            }
-        });
-
-        btn_QuitarProducto.setForeground(java.awt.Color.blue);
-        btn_QuitarProducto.setIcon(new javax.swing.ImageIcon(getClass().getResource("/sic/icons/DeleteProduct_16x16.png"))); // NOI18N
-        btn_QuitarProducto.setText("Quitar Producto (DEL)");
-        btn_QuitarProducto.setFocusable(false);
-        btn_QuitarProducto.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_QuitarProductoActionPerformed(evt);
-            }
-        });
-
-        txt_CodigoProducto.setFont(new java.awt.Font("DejaVu Sans", 0, 15)); // NOI18N
-        txt_CodigoProducto.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                txt_CodigoProductoActionPerformed(evt);
-            }
-        });
-
-        btn_BuscarPorCodigoProducto.setIcon(new javax.swing.ImageIcon(getClass().getResource("/sic/icons/16x16.png"))); // NOI18N
-        btn_BuscarPorCodigoProducto.setFocusable(false);
-        btn_BuscarPorCodigoProducto.setPreferredSize(new java.awt.Dimension(34, 28));
-        btn_BuscarPorCodigoProducto.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_BuscarPorCodigoProductoActionPerformed(evt);
-            }
-        });
-
-        tbtn_marcarDesmarcar.setFocusable(false);
-        tbtn_marcarDesmarcar.addChangeListener(new javax.swing.event.ChangeListener() {
-            public void stateChanged(javax.swing.event.ChangeEvent evt) {
-                tbtn_marcarDesmarcarStateChanged(evt);
-            }
-        });
-
         javax.swing.GroupLayout panelRenglonesLayout = new javax.swing.GroupLayout(panelRenglones);
         panelRenglones.setLayout(panelRenglonesLayout);
         panelRenglonesLayout.setHorizontalGroup(
             panelRenglonesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addComponent(sp_Resultado)
-            .addGroup(panelRenglonesLayout.createSequentialGroup()
-                .addComponent(tbtn_marcarDesmarcar, javax.swing.GroupLayout.PREFERRED_SIZE, 29, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(txt_CodigoProducto, javax.swing.GroupLayout.PREFERRED_SIZE, 149, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, 0)
-                .addComponent(btn_BuscarPorCodigoProducto, javax.swing.GroupLayout.PREFERRED_SIZE, 23, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(btn_BuscarProductos, javax.swing.GroupLayout.PREFERRED_SIZE, 173, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, 0)
-                .addComponent(btn_QuitarProducto)
-                .addGap(0, 0, Short.MAX_VALUE))
         );
-
-        panelRenglonesLayout.linkSize(javax.swing.SwingConstants.HORIZONTAL, new java.awt.Component[] {btn_BuscarProductos, btn_QuitarProducto});
-
         panelRenglonesLayout.setVerticalGroup(
             panelRenglonesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelRenglonesLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(panelRenglonesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
-                    .addComponent(txt_CodigoProducto, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(btn_BuscarPorCodigoProducto, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(tbtn_marcarDesmarcar, javax.swing.GroupLayout.PREFERRED_SIZE, 30, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(panelRenglonesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(btn_QuitarProducto, javax.swing.GroupLayout.PREFERRED_SIZE, 30, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(btn_BuscarProductos, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(sp_Resultado, javax.swing.GroupLayout.DEFAULT_SIZE, 237, Short.MAX_VALUE))
+            .addComponent(sp_Resultado, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 280, Short.MAX_VALUE)
         );
-
-        panelRenglonesLayout.linkSize(javax.swing.SwingConstants.VERTICAL, new java.awt.Component[] {btn_BuscarPorCodigoProducto, txt_CodigoProducto});
 
         lbl_Observaciones.setText("Observaciones:");
 
@@ -1097,15 +789,9 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
             }
         });
 
-        lbl_fechaDeVencimiento.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        lbl_fechaDeVencimiento.setText("Fecha de Vencimiento:");
-
-        dc_fechaVencimiento.setFocusable(false);
-
         lbl_TipoDeComprobante.setHorizontalAlignment(javax.swing.SwingConstants.RIGHT);
         lbl_TipoDeComprobante.setText("Tipo de Comprobante:");
 
-        cmb_TipoComprobante.setNextFocusableComponent(txt_CodigoProducto);
         cmb_TipoComprobante.addItemListener(new java.awt.event.ItemListener() {
             public void itemStateChanged(java.awt.event.ItemEvent evt) {
                 cmb_TipoComprobanteItemStateChanged(evt);
@@ -1117,99 +803,134 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
             }
         });
 
-        btn_NuevoCliente.setForeground(java.awt.Color.blue);
-        btn_NuevoCliente.setIcon(new javax.swing.ImageIcon(getClass().getResource("/sic/icons/AddClient_16x16.png"))); // NOI18N
-        btn_NuevoCliente.setText("Nuevo Cliente (F5)");
-        btn_NuevoCliente.setFocusable(false);
-        btn_NuevoCliente.setPreferredSize(new java.awt.Dimension(200, 30));
-        btn_NuevoCliente.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_NuevoClienteActionPerformed(evt);
-            }
-        });
-
-        btn_BuscarCliente.setForeground(java.awt.Color.blue);
-        btn_BuscarCliente.setIcon(new javax.swing.ImageIcon(getClass().getResource("/sic/icons/Client_16x16.png"))); // NOI18N
-        btn_BuscarCliente.setText("Buscar Cliente (F2)");
-        btn_BuscarCliente.setFocusable(false);
-        btn_BuscarCliente.setPreferredSize(new java.awt.Dimension(200, 30));
-        btn_BuscarCliente.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_BuscarClienteActionPerformed(evt);
-            }
-        });
-
         lblSeparadorDerecho.setHorizontalAlignment(javax.swing.SwingConstants.RIGHT);
 
-        btnModificarCliente.setForeground(java.awt.Color.blue);
-        btnModificarCliente.setIcon(new javax.swing.ImageIcon(getClass().getResource("/sic/icons/EditClient_16x16.png"))); // NOI18N
-        btnModificarCliente.setText("Modificar Cliente (F6)");
-        btnModificarCliente.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnModificarClienteActionPerformed(evt);
-            }
-        });
+        lblNombreCliente.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        lblNombreCliente.setText("Nombre:");
+
+        lblUbicacionCliente.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        lblUbicacionCliente.setText("Ubicación:");
+
+        lbl_IDFiscalCliente.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lbl_IDFiscalCliente.setText("CUIT o DNI:");
+
+        lbl_CondicionIVACliente.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        lbl_CondicionIVACliente.setText("Categoria IVA:");
+
+        txt_CondicionIVACliente.setEditable(false);
+        txt_CondicionIVACliente.setFocusable(false);
+
+        txtUbicacionCliente.setEditable(false);
+        txtUbicacionCliente.setFocusable(false);
+
+        txtNombreCliente.setEditable(false);
+        txtNombreCliente.setFocusable(false);
+
+        lblMontoCompraMinima.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        lblMontoCompraMinima.setText("Compra Mínima $:");
+
+        txtIdFiscalCliente.setEditable(false);
+        txtIdFiscalCliente.setFocusable(false);
+
+        txtMontoCompraMinima.setEditable(false);
+
+        javax.swing.GroupLayout panelClienteLayout = new javax.swing.GroupLayout(panelCliente);
+        panelCliente.setLayout(panelClienteLayout);
+        panelClienteLayout.setHorizontalGroup(
+            panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(panelClienteLayout.createSequentialGroup()
+                .addGap(0, 0, 0)
+                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addComponent(lblNombreCliente, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(lblUbicacionCliente, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(lbl_CondicionIVACliente))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelClienteLayout.createSequentialGroup()
+                        .addComponent(txt_CondicionIVACliente, javax.swing.GroupLayout.PREFERRED_SIZE, 581, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(lbl_IDFiscalCliente)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(txtIdFiscalCliente))
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelClienteLayout.createSequentialGroup()
+                        .addComponent(txtNombreCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 635, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(lblMontoCompraMinima)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(txtMontoCompraMinima))
+                    .addComponent(txtUbicacionCliente, javax.swing.GroupLayout.Alignment.TRAILING))
+                .addGap(0, 0, 0))
+        );
+
+        panelClienteLayout.linkSize(javax.swing.SwingConstants.HORIZONTAL, new java.awt.Component[] {txtNombreCliente, txt_CondicionIVACliente});
+
+        panelClienteLayout.setVerticalGroup(
+            panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(panelClienteLayout.createSequentialGroup()
+                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
+                    .addComponent(lblMontoCompraMinima)
+                    .addComponent(txtNombreCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblNombreCliente)
+                    .addComponent(txtMontoCompraMinima, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(txtUbicacionCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblUbicacionCliente))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(panelClienteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
+                    .addComponent(txtIdFiscalCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(txt_CondicionIVACliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lbl_IDFiscalCliente)
+                    .addComponent(lbl_CondicionIVACliente)))
+        );
+
+        panelClienteLayout.linkSize(javax.swing.SwingConstants.VERTICAL, new java.awt.Component[] {txtIdFiscalCliente, txtNombreCliente, txtUbicacionCliente, txt_CondicionIVACliente});
+
+        panelClienteLayout.linkSize(javax.swing.SwingConstants.VERTICAL, new java.awt.Component[] {lblMontoCompraMinima, lbl_IDFiscalCliente});
 
         javax.swing.GroupLayout panelEncabezadoLayout = new javax.swing.GroupLayout(panelEncabezado);
         panelEncabezado.setLayout(panelEncabezadoLayout);
         panelEncabezadoLayout.setHorizontalGroup(
             panelEncabezadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(panelEncabezadoLayout.createSequentialGroup()
+                .addContainerGap()
                 .addGroup(panelEncabezadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(panelEncabezadoLayout.createSequentialGroup()
-                        .addComponent(btn_NuevoCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 187, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(0, 0, 0)
-                        .addComponent(btn_BuscarCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 198, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(0, 0, 0)
-                        .addComponent(btnModificarCliente)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(lbl_fechaDeVencimiento)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(dc_fechaVencimiento, javax.swing.GroupLayout.PREFERRED_SIZE, 132, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(panelEncabezadoLayout.createSequentialGroup()
-                        .addContainerGap()
                         .addComponent(lblSeparadorIzquierdo, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                         .addComponent(lbl_TipoDeComprobante)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                         .addComponent(cmb_TipoComprobante, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(lblSeparadorDerecho, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                        .addComponent(lblSeparadorDerecho, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addComponent(panelCliente, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
                 .addContainerGap())
         );
-
-        panelEncabezadoLayout.linkSize(javax.swing.SwingConstants.HORIZONTAL, new java.awt.Component[] {btnModificarCliente, btn_BuscarCliente, btn_NuevoCliente});
-
         panelEncabezadoLayout.setVerticalGroup(
             panelEncabezadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.CENTER, panelEncabezadoLayout.createSequentialGroup()
-                .addGap(35, 35, 35)
-                .addComponent(lbl_fechaDeVencimiento, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addGap(6, 6, 6))
             .addGroup(panelEncabezadoLayout.createSequentialGroup()
                 .addGroup(panelEncabezadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(panelEncabezadoLayout.createSequentialGroup()
                         .addContainerGap()
                         .addGroup(panelEncabezadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
                             .addComponent(lbl_TipoDeComprobante)
-                            .addComponent(cmb_TipoComprobante, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGap(4, 4, 4)
-                        .addGroup(panelEncabezadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.CENTER)
-                            .addComponent(btn_BuscarCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 15, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(btnModificarCliente)
-                            .addComponent(btn_NuevoCliente, javax.swing.GroupLayout.PREFERRED_SIZE, 32, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                            .addComponent(cmb_TipoComprobante, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
                     .addGroup(panelEncabezadoLayout.createSequentialGroup()
                         .addGap(6, 6, 6)
                         .addGroup(panelEncabezadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                             .addComponent(lblSeparadorDerecho, javax.swing.GroupLayout.PREFERRED_SIZE, 25, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(lblSeparadorIzquierdo, javax.swing.GroupLayout.PREFERRED_SIZE, 25, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                    .addGroup(javax.swing.GroupLayout.Alignment.CENTER, panelEncabezadoLayout.createSequentialGroup()
-                        .addGap(39, 39, 39)
-                        .addComponent(dc_fechaVencimiento, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                            .addComponent(lblSeparadorIzquierdo, javax.swing.GroupLayout.PREFERRED_SIZE, 25, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(panelCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
-        panelEncabezadoLayout.linkSize(javax.swing.SwingConstants.VERTICAL, new java.awt.Component[] {btnModificarCliente, btn_BuscarCliente, btn_NuevoCliente});
+        tbtn_marcarDesmarcar.setFocusable(false);
+        tbtn_marcarDesmarcar.addChangeListener(new javax.swing.event.ChangeListener() {
+            public void stateChanged(javax.swing.event.ChangeEvent evt) {
+                tbtn_marcarDesmarcarStateChanged(evt);
+            }
+        });
 
         javax.swing.GroupLayout panelGeneralLayout = new javax.swing.GroupLayout(panelGeneral);
         panelGeneral.setLayout(panelGeneralLayout);
@@ -1220,14 +941,15 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
                 .addGroup(panelGeneralLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addComponent(panelRenglones, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addComponent(panelEncabezado, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(panelCliente, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addGroup(panelGeneralLayout.createSequentialGroup()
-                        .addComponent(panelResultados, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(0, 0, Short.MAX_VALUE))
                     .addGroup(panelGeneralLayout.createSequentialGroup()
                         .addComponent(panelObservaciones, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(btn_Continuar)))
+                        .addComponent(btn_Continuar))
+                    .addGroup(panelGeneralLayout.createSequentialGroup()
+                        .addGroup(panelGeneralLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(panelResultados, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(tbtn_marcarDesmarcar, javax.swing.GroupLayout.PREFERRED_SIZE, 29, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGap(0, 0, Short.MAX_VALUE)))
                 .addContainerGap())
         );
         panelGeneralLayout.setVerticalGroup(
@@ -1235,7 +957,7 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
             .addGroup(panelGeneralLayout.createSequentialGroup()
                 .addComponent(panelEncabezado, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(panelCliente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(tbtn_marcarDesmarcar, javax.swing.GroupLayout.PREFERRED_SIZE, 30, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(panelRenglones, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
@@ -1261,30 +983,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
-    private void btn_BuscarClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_BuscarClienteActionPerformed
-        BuscarClientesGUI buscarClientesGUI = new BuscarClientesGUI();
-        buscarClientesGUI.setModal(true);
-        buscarClientesGUI.setLocationRelativeTo(this);
-        buscarClientesGUI.setVisible(true);
-        if (buscarClientesGUI.getClienteSeleccionado() != null) {
-            this.cargarCliente(buscarClientesGUI.getClienteSeleccionado());
-            this.cargarTiposDeComprobantesDisponibles();
-            cmb_TipoComprobante.setSelectedIndex(0);
-        }
-    }//GEN-LAST:event_btn_BuscarClienteActionPerformed
-
-    private void btn_NuevoClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_NuevoClienteActionPerformed
-        DetalleClienteGUI gui_DetalleCliente = new DetalleClienteGUI();
-        gui_DetalleCliente.setModal(true);
-        gui_DetalleCliente.setLocationRelativeTo(this);
-        gui_DetalleCliente.setVisible(true);
-        if (gui_DetalleCliente.getClienteDadoDeAlta() != null) {
-            this.cargarCliente(gui_DetalleCliente.getClienteDadoDeAlta());
-            this.cargarTiposDeComprobantesDisponibles();
-            btnModificarCliente.setEnabled(true);
-        }
-    }//GEN-LAST:event_btn_NuevoClienteActionPerformed
-
     private void cmb_TipoComprobanteItemStateChanged(java.awt.event.ItemEvent evt) {//GEN-FIRST:event_cmb_TipoComprobanteItemStateChanged
         //para evitar que pase null cuando esta recargando el comboBox
         if (cmb_TipoComprobante.getSelectedItem() != null) {            
@@ -1292,14 +990,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
             this.recargarRenglones();
         }
     }//GEN-LAST:event_cmb_TipoComprobanteItemStateChanged
-
-    private void btn_BuscarPorCodigoProductoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_BuscarPorCodigoProductoActionPerformed
-        this.buscarProductoPorCodigo();
-    }//GEN-LAST:event_btn_BuscarPorCodigoProductoActionPerformed
-
-    private void txt_CodigoProductoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txt_CodigoProductoActionPerformed
-        this.buscarProductoPorCodigo();
-    }//GEN-LAST:event_txt_CodigoProductoActionPerformed
 
     private void btn_AddCommentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_AddCommentActionPerformed
         ObservacionesGUI observacionesGUI = new ObservacionesGUI(txt_Observaciones.getText());
@@ -1336,7 +1026,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
                     this.calcularResultados();
                     try {
                         cliente = RestClient.getRestTemplate().getForObject("/clientes/" + this.cliente.getIdCliente(), Cliente.class);
-                        List<ProductoFaltante> faltantes;
                         List<NuevoRenglonFactura> nuevosRenglones = new ArrayList<>();
                         this.renglonesFactura.forEach(renglon -> {
                             NuevoRenglonFactura nuevoRenglon = NuevoRenglonFactura.builder()
@@ -1345,21 +1034,11 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
                                     .build();
                             nuevosRenglones.add(nuevoRenglon);
                         });
-                        faltantes = new ArrayList<>();
-                        if (pedido != null) {
-                            faltantes = this.getProductosSinStockDisponible(nuevosRenglones);
-                        }
-                        if (faltantes.isEmpty()) {
-                            CerrarVentaGUI cerrarVentaGUI = new CerrarVentaGUI(this.construirNuevaFactura(), this.pedido, new BigDecimal(txt_Total.getValue().toString()), this.modeloTablaResultados);
-                            cerrarVentaGUI.setLocationRelativeTo(this);
-                            cerrarVentaGUI.setVisible(true);
-                            if (cerrarVentaGUI.isExito()) {
-                                this.dispose();
-                            }
-                        } else {
-                            ProductosFaltantesGUI productosFaltantesGUI = new ProductosFaltantesGUI(faltantes);
-                            productosFaltantesGUI.setLocationRelativeTo(this);
-                            productosFaltantesGUI.setVisible(true);
+                        CerrarOperacionGUI cerrarOperacionGUI = new CerrarOperacionGUI(this.construirNuevaFactura(), this.pedido, new BigDecimal(txt_Total.getValue().toString()), this.modeloTablaResultados);
+                        cerrarOperacionGUI.setLocationRelativeTo(this);
+                        cerrarOperacionGUI.setVisible(true);
+                        if (cerrarOperacionGUI.isExito()) {
+                            this.dispose();
                         }
                     } catch (RestClientResponseException ex) {
                         JOptionPane.showMessageDialog(this, ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
@@ -1376,37 +1055,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
                     "Aviso", JOptionPane.ERROR_MESSAGE);
         }
     }//GEN-LAST:event_btn_ContinuarActionPerformed
-
-    private void btn_QuitarProductoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_QuitarProductoActionPerformed
-        int[] indicesParaEliminar = Utilidades.getSelectedRowsModelIndices(tbl_Resultado);
-        EstadoRenglon[] estadoDeRenglones = new EstadoRenglon[renglonesFactura.size()];
-        for (int i = 0; i < tbl_Resultado.getRowCount(); i++) {
-            if (((boolean) tbl_Resultado.getValueAt(i, 0)) == true) {
-                estadoDeRenglones[i] = EstadoRenglon.MARCADO;
-            } else {
-                estadoDeRenglones[i] = EstadoRenglon.DESMARCADO;
-            }
-        }
-        for (int i = 0; i < indicesParaEliminar.length; i++) {
-            estadoDeRenglones[indicesParaEliminar[i]] = EstadoRenglon.ELIMINADO;
-        }
-        List<RenglonFactura> renglonesParaEliminar = new ArrayList<>();
-        for (int i = 0; i < indicesParaEliminar.length; i++) {
-            renglonesParaEliminar.add(renglonesFactura.get(indicesParaEliminar[i]));
-        }
-        renglonesFactura.removeAll(renglonesParaEliminar);
-        this.cargarRenglonesAlTable(estadoDeRenglones);
-        this.calcularResultados();
-    }//GEN-LAST:event_btn_QuitarProductoActionPerformed
-
-    private void btn_BuscarProductosActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_BuscarProductosActionPerformed
-        if (cliente != null) {
-            this.buscarProductoConVentanaAuxiliar();
-        } else {
-            JOptionPane.showMessageDialog(this, ResourceBundle.getBundle("Mensajes")
-                    .getString("mensaje_seleccionar_cliente"), "Error", JOptionPane.ERROR_MESSAGE);
-        }
-    }//GEN-LAST:event_btn_BuscarProductosActionPerformed
 
     private void tbl_ResultadoFocusGained(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_tbl_ResultadoFocusGained
         //Si no hay nada seleccionado y NO esta vacio el table, selecciona la primer fila
@@ -1477,27 +1125,12 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
             this.setSize(sizeInternalFrame);
             this.setColumnas();
             this.setMaximum(true);
-            cantidadMaximaRenglones = RestClient.getRestTemplate().getForObject("/configuraciones-sucursal/"
-                    + SucursalActiva.getInstance().getSucursal().getIdSucursal()
-                    + "/cantidad-renglones", Integer.class);
-            if (rolesDeUsuario.contains(Rol.ADMINISTRADOR)
-                    || rolesDeUsuario.contains(Rol.ENCARGADO)
-                    || rolesDeUsuario.contains(Rol.VENDEDOR)) {
-                if (this.existeClientePredeterminado()) {
-                    CuentaCorrienteCliente cuentaCorrienteClientePredeterminado
-                            = RestClient.getRestTemplate().getForObject("/cuentas-corriente/clientes/predeterminado", CuentaCorrienteCliente.class);
-                    this.cargarCliente(cuentaCorrienteClientePredeterminado.getCliente());
-                    this.btnModificarCliente.setEnabled(true);
-                }
-            }
             if (!this.existeFormaDePagoPredeterminada() || !this.existeTransportistaCargado()) {
                 this.dispose();
             }
             this.cargarTiposDeComprobantesDisponibles();
             if (this.pedido != null && this.pedido.getIdPedido() != 0) {
                 this.cargarPedidoParaFacturar();
-                btn_NuevoCliente.setEnabled(false);
-                btn_BuscarCliente.setEnabled(false);
                 this.calcularResultados();
             } else {
                 txt_Descuento_porcentaje.setValue(BigDecimal.ZERO);
@@ -1520,43 +1153,10 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
         }
     }//GEN-LAST:event_formInternalFrameOpened
 
-    private void btnModificarClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnModificarClienteActionPerformed
-        if (cliente != null) {
-            DetalleClienteGUI gui_DetalleCliente = new DetalleClienteGUI(cliente);
-            gui_DetalleCliente.setModal(true);
-            gui_DetalleCliente.setLocationRelativeTo(this);
-            gui_DetalleCliente.setVisible(true);
-            try {
-                this.cargarCliente(RestClient.getRestTemplate().getForObject("/clientes/" + this.cliente.getIdCliente(), Cliente.class));
-                this.recargarRenglones();
-            } catch (RestClientResponseException ex) {
-                JOptionPane.showMessageDialog(this, ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
-                this.dispose();
-            } catch (ResourceAccessException ex) {
-                LOGGER.error(ex.getMessage());
-                JOptionPane.showMessageDialog(this,
-                        ResourceBundle.getBundle("Mensajes").getString("mensaje_error_conexion"),
-                        "Error", JOptionPane.ERROR_MESSAGE);
-                this.dispose();
-            }
-        } else {
-            JOptionPane.showMessageDialog(this,
-                    ResourceBundle.getBundle("Mensajes").getString("mensaje_seleccionar_cliente"),
-                    "Error", JOptionPane.ERROR_MESSAGE);
-        }
-    }//GEN-LAST:event_btnModificarClienteActionPerformed
-
     // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btnModificarCliente;
     private javax.swing.JButton btn_AddComment;
-    private javax.swing.JButton btn_BuscarCliente;
-    private javax.swing.JButton btn_BuscarPorCodigoProducto;
-    private javax.swing.JButton btn_BuscarProductos;
     private javax.swing.JButton btn_Continuar;
-    private javax.swing.JButton btn_NuevoCliente;
-    private javax.swing.JButton btn_QuitarProducto;
     private javax.swing.JComboBox cmb_TipoComprobante;
-    private com.toedter.calendar.JDateChooser dc_fechaVencimiento;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JLabel lblMontoCompraMinima;
     private javax.swing.JLabel lblNombreCliente;
@@ -1575,7 +1175,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
     private javax.swing.JLabel lbl_SubTotalBruto;
     private javax.swing.JLabel lbl_TipoDeComprobante;
     private javax.swing.JLabel lbl_Total;
-    private javax.swing.JLabel lbl_fechaDeVencimiento;
     private javax.swing.JLabel lbl_recargoPorcentaje;
     private javax.swing.JPanel panelCliente;
     private javax.swing.JPanel panelEncabezado;
@@ -1590,7 +1189,6 @@ public class NuevaFacturaVentaGUI extends JInternalFrame {
     private javax.swing.JTextField txtMontoCompraMinima;
     private javax.swing.JTextField txtNombreCliente;
     private javax.swing.JTextField txtUbicacionCliente;
-    private javax.swing.JTextField txt_CodigoProducto;
     private javax.swing.JTextField txt_CondicionIVACliente;
     private javax.swing.JFormattedTextField txt_Descuento_neto;
     private javax.swing.JFormattedTextField txt_Descuento_porcentaje;
